tdf#130857 weld: Introduce weld::DrawingArea::signal_command
Instead of calling `m_aCommandHdl.Call` directly from the
subclasses, introduce a new method that does this, as is
done for other handlers.

Rename the existing GtkInstanceDrawingArea::signal_command
to GtkInstanceDrawingArea::signalCommand and let it call
the new base class method.
It is still needed as a public wrapper of the protected
base class method because it gets called from outside
of GtkInstanceDrawingArea itself.

This also prepares for deduplicating the CommandEvent
handling currently implemented in various weld::Widget
subclasses (weld::DrawingArea, weld::TreeView,
weld::IconView).

Change-Id: Ic556298ca5e96f934afbd5a1f74c0ac0e4d03743
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195413
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins