tdf#130857 weld: Move CommandEvent handling to weld::Widget base
So far, both weld::IconView and weld::DrawingArea
had a m_aCommandHdl member for handling/forwarding a
CommandEvent and associated methods to set it.

Unify/Deduplicate this in the weld::Widget base class.
In a first step, simply move the existing members there
and drop them in both subclasses.

This by itself does not yet imply that connecting a handler
to a different weld::Widget subclass will result in it getting
called with a corresponding CommandEvent when e.g. right-clicking
on that widget, because emitting such a CommandEventd is not yet
implemented in any other subclasses, but that can now be added
later as needed without duplicating the same logic in yet another
subclass.

Drop GtkInstanceIconView::signal_command, as it does the
same as the new weld::Widget::signal_command now.

Change-Id: I9367bdbb8be7994057667fab539aaf135751e063
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195414
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins