tdf#130857 weld: Reuse existing weld::Widget command logic for TreeView
Drop the weld::TreeView-specific logic for signalling
CommandEvent of type ContextMenu and reuse the existing
logic implemented in the weld::Widget base class instead,
which was moved there in previous commit

    Change-Id: I9367bdbb8be7994057667fab539aaf135751e063
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Dec 10 23:08:21 2025 +0100

        tdf#130857 weld: Move CommandEvent handling to weld::Widget base

Make weld::Widget::connect_command virtual to override
it in GtkInstanceTreeView, as was previously the case
for the now dropped weld::TreeView::connect_popup_menu.

The existing handlers previously connected using
weld::TreeView::connect_popup_menu already check whether the type
of the event is CommandEventId::ContextMenu, so no change in behavior
is expected.

In a quick test, the sample scenario of the template dialog as
described in

    commit 5d287676541d4f844b9b641a05f392d351f30d72
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Dec 9 20:13:07 2025 +0100

        tdf#130857 qt weld: Forward TreeView context menu event

still worked as expected with gtk3 and gen as well as with
an experimental branch for qt6 where support for using native
Qt widgets for that dialog is declared.

Change-Id: I13baf51eda2c9f0adc9c167ab50e920d23ada07e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195415
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>