tdf#130857 gtk weld: Forward keyboard-triggered context menu events, too
Previous commit

    Change-Id: Ie283a5f4975b7ba124abc69623d29c130c6ae2b6
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Dec 11 01:07:05 2025 +0100

        tdf#130857 gtk weld: Forward context menu events for all widgets

implemented forwarding context menu events for all widgets when
they are triggered using the mouse.

Implement the same for keyboard-triggered events, at least
for gtk3 (the GtkWidget::popup-menu event doesn't seem to exist
any more in GTK 4), for which this was previously implemented for the
weld::TreeView, weld::IconView and weld::DrawingArea subclasses.

Move the logic to the base class.

This results in the corresponding output when testing the
example as described in

        Change-Id: I07c5f136cc1671e1e0ca0ee3df11d2d9febc654a
        Author: Michael Weghorn <m.weghorn@posteo.de>
        Date:   Thu Dec 11 00:45:42 2025 +0100

            tdf#130857 qt weld: Forward context menu events for all widgets

, but now using the gtk3 VCL plugin and triggering the
context menu event by moving focus to the button and
pressing the context menu key on the keyboard instead
of right-clicking it using the mouse.

Change-Id: I919546761718384adb05fbaf8c3410999a0bba54
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195419
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>