Fix test checking metafile's coordinates
Respect the mapmode stored in the metafile, to enable it on all
platforms. On my Windows system, it was

  <mapmode mapunit="MapRelative" x="-1065" y="-284" scalex="(1/1)" scaley="(1/1)"/>

and on a Linux system, it was

  <mapmode mapunit="MapRelative" x="-284" y="-284" scalex="(1/1)" scaley="(1/1)"/>

which explained the difference.

Note that the metafile mapmode is obviously unstable. E.g., commit
I299e8f78b83c1d73402e548ea7f6cce1a2e5af0d causes the uncorrected test
fail on Linux with

 - Expected: 2306
 - Actual  : 4595
 - In <>, attribute 'x' of '//textarray[1]' incorrect value.

But curiously, running the test *individually*, like

 make CppunitTest_sw_odfexport4 CPPUNIT_TEST_NAME=testTdf78510

passes. So the change to mapmode is caused by some other test, which
is worrying, but that mapmode change is not addressed in this patch.
The fixed, more robust test will pass.

Change-Id: Ib004fd89aa03be38ab83d7ef70e0d5e791987298
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195421
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195504