svx: Move more SvxSearchCharSet::KeyInput logic to base class
Adjust the handling of the KEY_END case in the base
class to also cover the logic previously implemented
in SvxSearchCharSet::KeyInput.

SvxShowCharSet::getMaxCharCount returns
`mxFontCharMap->GetCharCount()`, so calling the former
is equivalent for the base class and covers the
logic previously only implemented in the subclass.

With that in place, only the default case is still
handled in SvxSearchCharSet::KeyInput itself, meaning
that `tmpSelected` will always be -1 and the logic
further down will never be triggered. Drop it and
simplify accordingly.

Change-Id: Ib9cf60eaba44f6745b4897ec1dfa53d581f3ae37
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195492
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins