fix nss/nspr build on intel macs
regression from aba5f8016965e42075e68faa17ddfd3b3213247d

nss' Darwin.mk uses uname -p to decide on the platform if CPU_ARCH is
not set, and that still returns i386 on intel macs – in the old
version that still worked since there was an additional ifdef USE_64
check that then set it to x86_64, but that has since been removed and
only the custom guards to not add additional -arch parameter to the
compiler flags remain, resulting in the built to attempt a 32bit and
that then fails when it comes to the assembly instructions:

mpi/mpi_sse2.s:35:5: error: instruction requires: Not 64-bit mode
    push %ebp
    ^
mpi/mpi_sse2.s:37:5: error: instruction requires: Not 64-bit mode
    push %edi
    ^
mpi/mpi_sse2.s:38:5: error: instruction requires: Not 64-bit mode
    push %esi
    ^
[…]

So don't rely on auto-detection and specify it explicitly as it was done
for the other variants already, ironically also already for building for
intel on arm. But the additional check whether it is building on arm is
not necessary, our CPUNAME is sufficient already and that now
superfluous bit can be removed.

Change-Id: Ic467b06beb9b73e2d177eedede886681aa668b15
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195439
Tested-by: Jenkins
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-by: Dan Williams <dcbw@libreoffice.org>