tdf#168594 Move logic to show context menu out of SvxCharView
Move the logic to show the context menu for one
of the recent or favorite characters in the
special characters popup and dialog from the
SvxCharView class to the SfxCharmapContainer
that manages all of the recent and favorite
characters, i.e. which owns the SvxCharView
instances.

No change in behavior intended or seen in a
quick test.

This is in preparation of switching away from
SvxCharView and using a weld::IconView for
the recent and favorite characters instead in an
upcoming commit.

Change-Id: Id4b5cee8abc68be3dfbdbdb26db5a675be2fd15f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195582
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins