tdf#168737 sw interdependent redlines, reject all: fix ignored format redlines
The bugdoc has two redlines: a format and a delete redline. Pressing
"reject all" in the manage changes dialog doesn't reject the format
redline.

Seems this happens because of code added in commit
9c4eef7d809ad7d283860c7b47b0f561aa240906 (tdf#52391 reject/clear
formatting of format-only changes, 2019-01-29), which argued that
rejection of format redlines isn't really implemented, so better if a
"mass reject all" doesn't reject them, unless only format redlines are
found in the document.  But in the meantime I worked on improving
support for rejecting format redlines, most recently commit
47cb60bc534104852a219426e3beb4c714b6cf07 (tdf#167761 sw format redline,
char props: implement ODF import, 2025-08-18) added ODF import/export of
these old direct character properties, so it now makes less sense to
filter out format redlines while rejecting all of them.  Also, the
matching toolbar button didn't ignore such format redlines.

Fix the problem by simplifying SwRedlineAcceptDlg::CallAcceptReject():
unless the redline data for the iterator says this item is disabled,
always add it to the list of redlines, similar to what the "reject the
selected redline(s)" case did already.

Note that the original 2019 commit was less confusing, as it did change
the label of the reject all button, but that was removed in commit
2e1a38ceb6866248ec30f6fe58cd3adc1b910eec (tdf#146893 Rework Manage
Changes dialog, 2022-01-24).

Change-Id: I69e7247d068cea2a25cd46ad2d5b41fccfa54919
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195772
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins