tdf#130857 tdf#168594 weld: Introduce weld::IconView::set_item_tooltip_text
Introduce a new weld::IconView method that allows to
directly set the tooltip for an item in the IconView
and implement the logic for all of the weld::IconView
implementations.

GtkIconView has a GtkIconView::tooltip-column property [1]
that is the index for the column in the underlying model
that holds the tooltip text.
For the gtk3/gtk4 implementation in GtkInstanceIconView,
that property can and will have to be set for the
GtkIconView in the .ui file - similar to what is
already the case for GtkIconView::pixbuf-column [2].

Prior to this commit, a tooltip could be set by setting a
handler using weld::IconView::connect_query_tooltip and
implementing logic to retrieve the tooltip text
in the handler.

Being able to set the tooltip directly simplifies
the handling where the tooltip for each item is fixed
and allows to set it right away, e.g. to the same text
that is already set as the item's accessible name
(using weld::IconView::set_item_accessible_name).

Upcoming commits will convert uses of
weld::IconView::connect_query_tooltip to use the
new API instead, see for example

    Change-Id: I0f41a5a4fd13a3296fe9dcd84ea71e7d2d153a97
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Dec 16 12:37:43 2025 +0100

        tdf#130857 Set fontwork item tooltips directly

[1] https://docs.gtk.org/gtk3/property.IconView.tooltip-column.html
[2] https://docs.gtk.org/gtk3/property.IconView.pixbuf-column.html

Change-Id: I56e8dde0b591073dea7f58190746ca0ba1fd8a4f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195711
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins