tdf#130857 Set fontwork item tooltips directly
Use the new API introduced in previous commit

    Change-Id: I56e8dde0b591073dea7f58190746ca0ba1fd8a4f
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Dec 16 12:10:48 2025 +0100

        tdf#130857 tdf#168594 weld: Introduce weld::IconView::set_item_tooltip_text

to directly set the tooltip text for each of the
items, just like the accessible name (which is using
the same string) is also set directly in
FontWorkGalleryDialog::fillFavorites.

For gtk3/gtk4, add a corresponding column in the
GtkIconView model and set GtkIconView::tooltip-column
accordingly, as described in the above-mentioned commit.

This also gets rid of the need to remember the titles for
the items in a map.

Simplify the loop a bit by using the index and adding 1
once instead of using index+1 and subtracting 1 everywhere
else to get the index again.

The dialog can be triggered using "Insert" -> "Fontwork..."
in Writer.

Change-Id: I0f41a5a4fd13a3296fe9dcd84ea71e7d2d153a97
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195712
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins