tdf#130857 Set iconview tooltips in "Borders" tab directly
Use the new API introduced in previous commit

    Change-Id: I56e8dde0b591073dea7f58190746ca0ba1fd8a4f
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Dec 16 12:10:48 2025 +0100

        tdf#130857 tdf#168594 weld: Introduce weld::IconView::set_item_tooltip_text

to directly set the tooltip text for each of the
IconView items, just like the accessible name (which is using
the same string) is also set directly already.

For gtk3/gtk4, add a corresponding column in the
GtkIconView model and set GtkIconView::tooltip-column
accordingly, as described in the above-mentioned commit.

This is e.g. used for the "Borders" tab page in Writer's
"Format" -> "Page Style" dialog.

Change-Id: Idf772d2156bff9f220e304acaed8ddef8780d7fe
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195713
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins