tdf#130857 sd: Set iconview tooltips in LayoutMenu directly
Use the new API introduced in previous commit

    Change-Id: I56e8dde0b591073dea7f58190746ca0ba1fd8a4f
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Dec 16 12:10:48 2025 +0100

        tdf#130857 tdf#168594 weld: Introduce weld::IconView::set_item_tooltip_text

to directly set the tooltip text for each of the
IconView items.

This means it's no longer necessary to remember a mapping
between the ID and the layout name, so drop the corresponding
logic.

For gtk3/gtk4, add a corresponding column in the
GtkIconView model and set GtkIconView::tooltip-column
accordingly, as described in the above-mentioned commit.

This is used in the Impress sidebar, "Properties" panel.

Change-Id: I3840d4b7122a423000f3edb6f5ee64ac8118b39b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195714
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins