sd a11y: Set a11y names for "Layout" icon view items
For the layout items in the Impress sidebar, "Properties" panel,
section "Layout", set the layout names already used for the
tooltips also as the accessible names.

This makes e.g. the Orca screen reader on Linux with
the gtk3 VCL plugin announce those names when focus
moves between the items, instead of just saying "Icon".

Change-Id: I069673ebff142b0b321ffe9fca72ab5da3a459af
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195715
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>