sw: fix AutoFormat applying TextBody style
The problem is that BuildText() applies the "Text body" style, and it's
now called even in the !m_aFlags.bAFormatByInput case (previously
it (or other Build* functions that also set styles) was only called when
the user runs the AutoFormat menu item).

Not calling it at all lets the testTdf139922 fail, as it also calls
AutoCorrect(), which this test checks.

So let's try 2 things:
1. only call BuildText when it's explicit AutoFormat
2. skip setting the text body style if the new config that was introduced
   by the same commit is disabled.

(regression from commit 5bd1efb830400e1394d45f122faf7537e4451f33)

Change-Id: Ib7d0bad479fe52c013a1b70633de7cb01db047f5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195733
Tested-by: allotropia jenkins <jenkins@allotropia.de>
Reviewed-by: Thorsten Behrens <thorsten.behrens@collabora.com>