tdf#168594 Move tooltip text and a11y name logic out of SvxCharView
Move the logic to generate the tooltip text and
the accessible name from the SvxCharView class to
the SfxCharmapContainer that manages all of the recent
and favorite characters, i.e. which owns the SvxCharView
instances.
(Make the relevant SvxCharView methods static methods
in SfxCharmapContainer instead.)

No change in behavior intended or seen in a
quick test.

This is in preparation of switching away from
SvxCharView and using a weld::IconView for
the recent and favorite characters instead in an
upcoming commit.

With previous commit

    Change-Id: I56e8dde0b591073dea7f58190746ca0ba1fd8a4f
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Dec 16 12:10:48 2025 +0100

        tdf#130857 tdf#168594 weld: Introduce weld::IconView::set_item_tooltip_text

in place, weld::IconView now has API to set both, the
accessible name and the tooltip for an item directly which
should make porting the logic newly added to
SfxCharmapContainer::updateCharControl in this commit
straight-forward when switching to weld::IconView in
the future.

Change-Id: Ica33d45f7b9dbb6840ad67b725c7ed98ff505312
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195736
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins