weld: Move weld::EntryTreeView to its own header
Move the class out of weld.hxx into its own header, in
order to make it easier to keep an overview and to avoid
having to recompile ~all UI code when the class is
modified.

This also prepares for moving weld::TreeView
to its own header in an upcoming commit.

Change-Id: I7f525d435d40ab0e1b5d8cff316ec7b3c3728e70
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195780
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>