sd: Default drawing style to use ComplexColor for fill and line
This changes the default drawing style to use theme colors for
fill and line colors that will be used by by shapes. The use of
theme color is needed so that changing the document theme will
have an effect by default. This also means the default color
will be different than before - it will use the colors defined
by the default theme used by documents (LibreOffice theme).

Additionally it increases the default line size to make the
shapes more atttractive.

Test testThemeShapeInsert has been re-added and changed to pass.
Other tests have been fixed or changed to set to use the default
theme by force, which uses the old fill color for accent{1,2,3}
and old line color for accent{4,5,6}. With this the tests don't
need to be changed if they check the FillColor, which is set to
a default color (from the style).

Change-Id: Ibb16961dec79232603619efe040db70d90e07cee
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195787
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins