tdf#130857 qt weld: Introduce QtInstanceItemView
This is the implementation for the abstract
weld::ItemView class introduced in

    commit 94aee0cd3d6cfc2bdc8375f9e1938a92c15eae1f
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Dec 17 12:14:41 2025 +0100

        weld: Introduce weld::ItemView as IconView/TreeView base

Having an equivalent for the Qt-based weld implementation
will allow sharing code between QtInstanceIconView
and QtInstanceTreeView.

In an initial step, deduplicate the weld::ItemView::do_clear
implementations by moving the QtInstanceTreeView one to
the new base class and dropping the implementations in the
subclasses.

Change-Id: I91edd89743afc5cfa53d3b8343a3a69ab0a0c51b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195795
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>