tdf#130857 qt weld: Deduplicate IconView/TreeView iter logic
Now that

    Change-Id: I91edd89743afc5cfa53d3b8343a3a69ab0a0c51b
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Dec 17 14:28:38 2025 +0100

        tdf#130857 qt weld: Introduce QtInstanceItemView

introduced a shared base class, move the QtInstanceTreeView
logic to convert between position/index, model index and
iterators to the new QtInstanceItemView base class.

Reuse that existing logic in QtInstanceIconView as well,
and drop the logic from there that is comparable, but
only handles one dimension.
(The implemenation in QtInstanceItemView uses a default
column index of 0, which is fine.)

Adjust QtInstanceIconView::get_id to follow the same
logic as implemented in QtInstanceTreeView: In the variant
called with an index, call the variant taking an iterator,
not the other way around.

Change-Id: I045a7ca336f68c1128605aadef63b0303839a02c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195796
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>