vcl weld: Introduce SalInstanceItemView
This is the implementation for the abstract
weld::ItemView class introduced in

    commit 94aee0cd3d6cfc2bdc8375f9e1938a92c15eae1f
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Dec 17 12:14:41 2025 +0100

        weld: Introduce weld::ItemView as IconView/TreeView base

Having an equivalent for the vcl-based weld implementation
will allow sharing code between SalInstanceIconView
and SalInstanceTreeView.

In an initial step, move the m_aUserData member that both
subclasses have to the base class and deduplicate the
weld::ItemView::do_clear implementations by implementing the
logic in the base class and dropping the equivalent overrides
in both subclasses.

Change-Id: I6f35e2cac1c2f1851590307699631c2b9598f9eb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195797
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins