gtk weld: Introduce GtkInstanceItemView
This is the implementation for the abstract
weld::ItemView class introduced in

    commit 94aee0cd3d6cfc2bdc8375f9e1938a92c15eae1f
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Dec 17 12:14:41 2025 +0100

        weld: Introduce weld::ItemView as IconView/TreeView base

Having an equivalent for the GTK-based weld implementation
will allow sharing code between GtkInstanceIconView
and GtkInstanceTreeView.

In an initial step, deduplicate the weld::ItemView::make_iterator
implementations by implementing that method in the new base class
and dropping the two equivalent subclass overrides.

Change-Id: Id3f87567fc418e6b69eaf92b5b31a7e9ef59814b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195799
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>