weld: Add method to get ItemView iter for position
Introduce weld::ItemView::get_iterator that returns
a weld::TreeIter pointing to the item at the
given position/index if the position is valid, otherwise
a unique_ptr not owning anything.

This can be used in upcoming commits to unify the
iter-based and index-based methods in the weld::ItemView
subclasses (weld::TreeView, weld::IconView) or easily
provide both variants without having to implement
the whole logic twice in every implementation.
(If the gtk, vcl and qt implementations provide an
iter-based implementation, the abstract base class
can implement the index-based one itself, by converting
the index to an iter and calling the variant taking
an iter.)

Change-Id: I6c09044ba50556dd57fdc7bce399ba334efeada9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195801
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>