weld: Add weld::IconView::get_rect taking iter param
Change the virtual weld::IconView::get_rect param from
an index/position to a weld::TreeIter and implement
that one in the toolkit-specific implementations.

For the existing variant taking an index, implement
it directly in the weld::IconView base class instead,
by converting the index to a iterator and calling
the variant taking an iterator.

This makes use of the method introduced in previous commit

    Change-Id: I6c09044ba50556dd57fdc7bce399ba334efeada9
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Dec 17 16:47:31 2025 +0100

        weld: Add method to get ItemView iter for position

, see also that commit's commit message for more background.

Other methods could be converted/deduplicated in a similar
way, e.g. various weld::TreeView ones that currently have custom
toolkit-specific implementations for both variants.

No change intended or seen e.g. for the following use
case that makes use of the index-based method
(see LayoutMenu::MousePressHdl):

* start Impress
* open the "Properties" sidebar panel
* in the "Layouts" section, right-click one of
  the layouts to open the context menu for the
  given layout

The new iterator-based method will be used in an
upcoming commit to implement logic to also allow
opening the menu using the keyboard (by pressing
the context menu key when one of the IconView items
has focus).

Change-Id: Icac5e42196e8c490cc95131508d427e87d1bcb57
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195802
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins