sd a11y: Allow opening "Layout" context menu using keyboard
So far, the context menu for the currently selected
layout in the Impress "Layout" sidebar section could
only be opened using the mouse, not the keyboard.

Make both cases work, by implementing a command
handler that handles the CommandId::ContextMenu
case instead of a mouse event handler.

If the context menu was opened using the keyboard,
use the center of the currently selected item's
rect as the position for the context menu.

To trigger this:

* start Impress
* open the "Properties" sidebar panel
* in the "Layouts" section, move between layouts
  using the keyboard
* press the context menu key (or Shift+F10) to
  open the context menu for the currently selected
  item

This makes use of the new method introduced in
previous commit

    Change-Id: Icac5e42196e8c490cc95131508d427e87d1bcb57
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Dec 17 17:15:08 2025 +0100

        weld: Add weld::IconView::get_rect taking iter param

Change-Id: Iaae1b413b3b237f22c05d389f5960e1ae72a99e8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195803
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>