tdf#170012 sc drwlayer: avoid crash on ODS save
This fixes a crash since 7.1
commit 1f0b3c7a40edfa81bbc7a58d123a6a2dfd83e4ca

Instead of trying to save as a broken SCA_CELL_RESIZE
it will save as a SCA_PAGE anchored object.

The crash was occuring when saving to ODS format.
It would loop forever because of the invalid address
that GetNonRotatedObjData got during RecalcPos
when it had been created but never "initialized".

(An earlier patchset tried to force initialization
and assert that, but I got tons of asserts in unit tests
because of detfunc's bCreate without
ScDrawLayer::InitializeCellAnchoredObj).
Plus, doing initialization correctly
sounds incomprehensible based on a code read.

[Unit test removed because it depends on xlsx
importing correctly - which doesn't work in 25.8.]

Change-Id: Ief717c8742d712bf2edec90f492fd258f9fd7741
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195807
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
(cherry picked from commit cbd9d23b7b953fdf53d66aaa072bab17db7905d3)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195861
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195873
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>