tdf#130857 weld: Restrict TreeView text alignment to left/center/right
So far, a floating point value was used, but for the
vcl/SalInstanceTreeView implementation, this was already
mapped to an alignment of only left/center/right.

Qt also only supports these via QStandardItem::setTextAlignment [1].

Restrict the API to the same, and simply map that to 0, 0.5 and 1.0
again in the GtkInstanceTreeView implementation.

So far, CustomAnimationPane is the only place where this API
is used, and it uses a value of 0.5 (for centered text).

No change in behavior intended or seen in a quick test of
that pane with gtk3 or gen:

* start Impress
* select the title textbox
* open the "Animation" sidebar panel
* click the "Add" button

Some texts in the treeview in the "Effects" section
are bold and centered.

[1] https://doc.qt.io/qt-6/qstandarditem.html#setTextAlignment

Change-Id: Ibe184c9ae590a63511b61569035f98e4c6a7fa39
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195819
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins