Related: tdf#89365 sd UI, from numbering to bullet: fix defaults
Open the bugdoc, start text edit on the only slide, in the only shape.
Use the toolbar button to change from no numbering to bullet, some very
small bullet with no spacing between the bullet and the text shows up.

This happens since commit f60ee00edcc5b0fdee5227bb695448119cddb013
(tdf#89365 sd UI: fix transitioning from a numbered list to a bulleted
list, 2025-11-04), previously the numbering to bullet transitioning was
broken and you had to click on the "turn on bullets" button twice. A
side effect of the old two-step transitioning was that no formatting
from the old numbering was kept, so better defaults were used (but the
numbering level was lost).

Keep the old use-case fixed and fix the new use-case by looking at how
this works on master pages: there TextObjectBar::ExecuteImpl()'s
FN_NUM_BULLET_ON block calls
SdStyleSheetPool::setDefaultOutlineNumberFormatBulletAndIndent() before
setting the number format for a specific level of the numbering rule. Do
the same for non-master pages in
FuBulletAndPosition::SetCurrentBulletsNumbering(): if we transition to a
bullet, then use the same defaults in the non-master-page case, too.

Note that turning bullets on or off as direct formatting is not ideal: a
better style is to assign bullets (or no bullets) to different outline
levels in the master page, and then just change the list level of
paragraphs in the outliner shape, which avoids this sort of problem in
the first place.

Change-Id: I4227988593baeaa91cc21525fb895c3acaf3afd7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195839
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins
(cherry picked from commit fe9fab9702613b1a5d192821d8a620aa527234b7)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195885
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>