tdf#130857 weld: Move signal blocking to ItemView::{un,}select_all
This is similar to

    commit 039f89a0b91ee03b9357b2082bfe17e8769e1573
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Oct 17 22:11:02 2025 +0200

        tdf#130857 weld: Move signal blocking to TextView::select_region

, but for a different method:

The GTK and vcl/SalInstance implementations of
select_all() and unselect_all() implicitly block
signals by calling unselect() and select() with a
special index of -1. Then, weld::ItemView::unselect
and weld::ItemView::select would take care of blocking
signals while calling do_unselect() and do_select().

The QtInstanceItemView implementations on the other
hand implement the logic directly, so would not
block the signals.

This also prepares for implementing the logic
to (un)select all directly in the GTK and vcl
implementations in an upcoming commit, in order
to prepare for further refactoring.

Change-Id: Ie0453f63715b05339df530975faec095b472fedb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/195866
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>